requirements.txt
bs4==0.0.2
pyspellchecker==0.8.1
python-dotenv==1.0.1
spellchecker==0.4
